Game implementation improvements for early playtesting
- Improved seeking
- Improved catching reliability
We now size the colliders for the detectors as well as the catchers, but this will require more work, as it isn't quite accurate just yet, but it works for now.
- Loading song data from game directory now
The songs themselves are still temporarily embedded into the game assets.
- Fixed track assignments
Moggsong files don't have the correct track assignments - they are meant to be read from the MIDI.