Add ability to use fudge scale in pos -> millisecond conversions
This is used when genereating the catching slopMs and slopPos values.
Here, we want to scale the slop time to the fudge factor, because you move faster with a lower scale.
Without (scale), the time is converted back to its original value, as if there was no fudge scaling in place.