Optimization of track streaming (recycling)
We now recycle measures and notes, so that we don't have to instantiate brand new objects.
This is still not the most ideal in terms of performance, as the meshes still have to be re-deformed, which does a bunch of GC.Alloc()s.
Also, rendering / our ClippingPlane shaders are horrible!